@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import {ActivatedRoute, Router} from "@angular/router";
-import {FormArray, FormControl, FormGroup, Validators} from "@angular/forms";
+import {UntypedFormArray, UntypedFormControl, UntypedFormGroup, Validators} from "@angular/forms";
 import {RecipeService} from "../recipe.service";
 
 @Component({
@@ -11,12 +11,12 @@ import {RecipeService} from "../recipe.service";
 export class RecipeEditComponent implements OnInit {
   id: number;
   editMode = false;
-  recipeForm: FormGroup;
+  recipeForm: UntypedFormGroup;
 
   constructor(private route: ActivatedRoute, private recipeService: RecipeService, private router: Router) { }
 
   get controls() {
-    return (<FormArray>this.recipeForm.get('ingredients')).controls;
+    return (<UntypedFormArray>this.recipeForm.get('ingredients')).controls;
   }
 
   ngOnInit(): void {
@@ -32,7 +32,7 @@ export class RecipeEditComponent implements OnInit {
     let recipeName = '';
     let recipeImagePath = '';
     let recipeDescription = '';
-    let recipeIngredients = new FormArray([]);
+    let recipeIngredients = new UntypedFormArray([]);
 
     if (this.editMode) {
       const recipe = this.recipeService.getRecipe(this.id);
@@ -43,9 +43,9 @@ export class RecipeEditComponent implements OnInit {
       if (recipe['ingredients']) {
         console.log(recipe.ingredients);
         for (let ingredient of recipe.ingredients) {
-          recipeIngredients.push(new FormGroup({
-            'name': new FormControl(ingredient.name, Validators.required),
-            'amount': new FormControl(ingredient.amount, [
+          recipeIngredients.push(new UntypedFormGroup({
+            'name': new UntypedFormControl(ingredient.name, Validators.required),
+            'amount': new UntypedFormControl(ingredient.amount, [
               Validators.required,
               Validators.pattern('^[1-9]+[0-9]*$')
             ])
@@ -54,10 +54,10 @@ export class RecipeEditComponent implements OnInit {
       }
     }
 
-    this.recipeForm = new FormGroup({
-      'name': new FormControl(recipeName, Validators.required),
-      'imagePath': new FormControl(recipeImagePath, Validators.required),
-      'description': new FormControl(recipeDescription, Validators.required),
+    this.recipeForm = new UntypedFormGroup({
+      'name': new UntypedFormControl(recipeName, Validators.required),
+      'imagePath': new UntypedFormControl(recipeImagePath, Validators.required),
+      'description': new UntypedFormControl(recipeDescription, Validators.required),
       'ingredients': recipeIngredients
     });
   }
@@ -72,9 +72,9 @@ export class RecipeEditComponent implements OnInit {
   }
 
   onAddIngredient() {
-    (<FormArray>this.recipeForm.get('ingredients')).push(new FormGroup({
-      name: new FormControl(null, Validators.required),
-      amount: new FormControl(null, [
+    (<UntypedFormArray>this.recipeForm.get('ingredients')).push(new UntypedFormGroup({
+      name: new UntypedFormControl(null, Validators.required),
+      amount: new UntypedFormControl(null, [
         Validators.required,
         Validators.pattern('^[1-9]+[0-9]*$')
       ])
@@ -86,6 +86,6 @@ export class RecipeEditComponent implements OnInit {
   }
 
   onDeleteIngredient(index: number) {
-    (<FormArray>this.recipeForm.get('ingredients')).removeAt(index);
+    (<UntypedFormArray>this.recipeForm.get('ingredients')).removeAt(index);
   }
 }
